+++ Samples~/Bootstrap MainLoop/Scripts/Configuration.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace ReplaceMeWithYourProjectsNamespace
+{
+    internal class Configuration : IConfiguration
+    {
+        public float TimeScaleFactor { get; set; }
+    }
+}