+++ Samples~/Bootstrap MainLoop/Scripts/IRuntimeData.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace ReplaceMeWithYourProjectsNamespace
+{
+    internal interface IRuntimeData
+    {
+        float SimulationTimeSec { get; set; }
+    }
+}