+++ Samples~/Bootstrap MainLoop/Scripts/ITimeLapse.cs
@@ -1,0 +1,7 @@
+﻿namespace ReplaceMeWithYourProjectsNamespace
+{
+    public interface ITimeLapse
+    {
+        float ToSimulationTime( float realTimeSec );
+    }
+}