+++ CHANGELOG.md
@@ -1,4 +1,6 @@
+## [1.2.0] - 2019-12-07
+Added Bootstrap MainLoop Example
 ## [1.1.0] - 2019-12-07
-Implemented support for RegisterInstance and SingleInstance as well constructor parameter resolution.
+Implemented support for RegisterInstance and SingleInstance as well constructor parameter resolution
 ## [1.0.0] - 2019-12-05
-Initial Release.
+Initial Release